fix(ci): format merged test code and allow test-only pedantic lints
Merged contributor test PRs introduced rustfmt drift and clippy pedantic
lints (empty lines after attrs/doc comments, absurd-extreme comparisons,
etc.) in the test modules. Reformatted and scoped the allows to the test
module; no contract/wasm code affected.

+++ escrow/src/tests/attestations.rs
@@ -779,5 +779,5 @@ fn test_revoked_digests_view_caps_limit() {
         client.revoke_attestation_digest(&(i as u32));
     }
     let page = client.get_revoked_attestation_digests(&0, &100);
-    assert_eq!(page.len(), crate::MAX_ATTESTATION_READ_PAGE as u32);
+    assert_eq!(page.len(), crate::MAX_ATTESTATION_READ_PAGE);
 }
